Refactor tests: Extract helpers.js
